refactor: kebab-case filenames
`unicorn/filename-case` is on at kebabCase in the shared preset, so 20
files move. `magic-kebab --write` did the renames and the 34 specifier
rewrites; it was pointed at examples/kitchen-sink/tsconfig.json because
the default lookup only reads a tsconfig at the repo root and would have
left the four `@/components/...` imports pointing at the old names.

Three references the codemod does not look at were fixed by hand:
`app.config.ts` names the expo config plugin by path,
`pan-gesture-surface.test.ts` passes module paths as arguments to a
`load()` helper rather than to `require` directly, and a handful of doc
comments name their sibling test files.

Directories keep their casing — the rule only inspects basenames, and
renaming `MagicModalPortal/` or `panGesture/` is a separate change with
its own import churn.
