+++ packages/modal/src/components/magic-modal.swipe.v2.test.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { act } from "@testing-library/react-native";
 
-import type { LegacyPanGesture } from "./panGesture/PanGestureSurface.v2";
+import type { LegacyPanGesture } from "./panGesture/pan-gesture-surface.v2";
 import { MagicModalHideReason } from "../constants/types";
-import { magicModal } from "../utils/magicModalHandler";
+import { magicModal } from "../utils/magic-modal-handler";
 import {
   baseEvent,
   directions,
@@ -13,11 +13,11 @@ import {
   showModal,
   slowVelocity,
   waitForHide,
-} from "./MagicModal.swipe.harness";
+} from "./magic-modal.swipe.harness";
 
 /**
  * The gesture-handler 2.x half of the swipe coverage. Its 3.x counterpart is
- * `MagicModal.swipe.v3.test.tsx`; the assertions run in the same order so the
+ * `magic-modal.swipe.v3.test.tsx`; the assertions run in the same order so the
  * two read side by side.
  *
  * 2.x has no `usePanGesture`, so `MagicModal` builds a `Gesture.Pan()` instead.
