+++ packages/modal/src/components/magic-modal.swipe.v3.test.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { act } from "@testing-library/react-native";
 
-import type { PanGestureConfigCompat } from "./panGesture/gestureHandlerCompat";
+import type { PanGestureConfigCompat } from "./panGesture/gesture-handler-compat";
 import { MagicModalHideReason } from "../constants/types";
-import { magicModal } from "../utils/magicModalHandler";
+import { magicModal } from "../utils/magic-modal-handler";
 import {
   baseEvent,
   directions,
@@ -13,7 +13,7 @@ import {
   showModal,
   slowVelocity,
   waitForHide,
-} from "./MagicModal.swipe.harness";
+} from "./magic-modal.swipe.harness";
 
 /**
  * The swipe gesture is the one part of the modal the public API can't drive: it
@@ -30,7 +30,7 @@ import {
  * Defining `usePanGesture` in the mock also pins which surface renders:
  * `hasPanGestureHook` reads it off the module, so this suite exercises the
  * hook-based path whichever gesture-handler is installed. Its 2.x counterpart is
- * `MagicModal.swipe.v2.test.tsx`.
+ * `magic-modal.swipe.v2.test.tsx`.
  */
 jest.mock("react-native-gesture-handler", () => {
   const actual = jest.requireActual<Record<string, unknown>>(
