+++ packages/modal/src/components/panGesture/pan-gesture-surface.v3.tsx
@@ -1,9 +1,9 @@
 import React, { useMemo } from "react";
 import { GestureDetector } from "react-native-gesture-handler";
 
-import type { PanGestureConfigCompat } from "./gestureHandlerCompat";
-import type { PanGestureSurfaceProps } from "./PanGestureSurface.types";
-import { usePanGestureCompat } from "./gestureHandlerCompat";
+import type { PanGestureConfigCompat } from "./gesture-handler-compat";
+import type { PanGestureSurfaceProps } from "./pan-gesture-surface.types";
+import { usePanGestureCompat } from "./gesture-handler-compat";
 
 /**
  * Swipe surface for gesture-handler 3.x, built on the `usePanGesture` hook.
