+++ packages/modal/src/constants/default-config.ts
