+++ packages/modal/src/utils/magic-modal-handler.ts
